check for old images and processes on removing package

+++ rpm_packaging/before-remove.sh
@@ -4,6 +4,10 @@ systemctl disable goldstone-server
 
 # remove stopped containers
 echo "Removing containers..."
-docker rm -f $(docker ps -a | grep goldstone | awk '{print $1}' | xargs) || /bin/true
+if [ `docker ps -a | grep goldstone | wc -l` -gt 0 ]; then
+  docker rm -f $(docker ps -a | grep goldstone | awk '{print $1}' | xargs) || /bin/true
+fi
 echo "Removing images..."
-docker rmi -f $(docker images | grep goldstone | awk '{print $3}' | xargs) || /bin/true
+if [ `docker images -a | grep goldstone | wc -l` -gt 0 ]; then
+  docker rmi -f $(docker images | grep goldstone | awk '{print $3}' | xargs) || /bin/true
+fi
